move validations to configuration
